videos: rename components
