+++ web/mocksite/index.pug
@@ -27,7 +27,7 @@ block main
 
 	section
 		h2 videos
-		xiome-video-link
+		xiome-video-hosting
 		xiome-video-display
 
 	//- section
