+++ web/mocksite/videos.pug
@@ -5,7 +5,7 @@ block main
 
 	section
 		h2 videos
-		xiome-video-link
+		xiome-video-hosting
 		xiome-video-display
 
 	section
